requirements.txt
tqdm
label-studio
numpy
ultralytics
Pillow
pygetwindow
pyautogui
opencv-python
python-dotenv
tk
